Skyscope-Sentinel-AI-STUDIO-V1/requirements.txt
# Core AI and Agent Framework
openai-unofficial==0.1.2
langchain>=0.0.267
crewai>=0.28.0
swarms>=0.3.0
autogen>=0.2.0
agentverse>=0.2.0
instructor>=0.4.5
pydantic>=2.4.0
chromadb>=0.4.15
sentence-transformers>=2.2.2
llama-index>=0.8.54
guidance>=0.1.0
anthropic>=0.5.0
cohere-api>=4.32.0
google-generativeai>=0.3.0
transformers>=4.35.0
accelerate>=0.23.0
bitsandbytes>=0.41.0
optimum>=1.13.0
huggingface-hub>=0.17.3

# UI and Visualization
streamlit>=1.28.0
streamlit-extras>=0.3.0
streamlit-ace>=0.1.1
streamlit-chat>=0.1.1
streamlit-option-menu>=0.3.6
streamlit-webrtc>=0.47.0
streamlit-elements>=0.1.0
streamlit-bokeh-events>=0.1.2
bokeh>=3.2.2
plotly>=5.17.0
altair>=5.1.2
pandas>=2.1.1
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
pillow>=10.0.1
gradio>=4.1.1

# Web Automation and Scraping
playwright>=1.39.0
selenium>=4.14.0
beautifulsoup4>=4.12.2
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.8.5
scrapy>=2.11.0
newspaper3k>=0.2.8
feedparser>=6.0.10
pyppeteer>=1.0.2

# File System and Terminal Access
pyautogui>=0.9.54
pynput>=1.7.6
pyperclip>=1.8.2
python-dotenv>=1.0.0
watchdog>=3.0.0
psutil>=5.9.5
paramiko>=3.3.1
pysftp>=0.2.9
scp>=0.14.5

# Cryptocurrency and Finance
ccxt>=4.0.0
web3>=6.10.0
python-binance>=1.0.19
coinbase>=2.1.0
pycoingecko>=3.1.0
ta>=0.10.2
pandas-ta>=0.3.14b0
yfinance>=0.2.31
alpaca-trade-api>=3.0.2
robin-stocks>=2.1.0

# PDF and Document Processing
reportlab>=4.0.4
pdfminer.six>=20221105
PyPDF2>=3.0.1
python-docx>=0.8.11
openpyxl>=3.1.2
python-pptx>=0.6.21
docx2pdf>=0.1.8
weasyprint>=60.1
xhtml2pdf>=0.2.11
markdown>=3.5
jinja2>=3.1.2

# Natural Language Processing
spacy>=3.7.1
nltk>=3.8.1
textblob>=0.17.1
gensim>=4.3.2
keybert>=0.8.3
sumy>=0.11.0
lexrank>=0.1.0
polyglot>=16.7.4
fasttext>=0.9.2

# Image and Audio Processing
opencv-python>=4.8.1.78
scikit-image>=0.22.0
imageio>=2.31.5
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
speechrecognition>=3.10.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Security and Authentication
cryptography>=41.0.4
pyotp>=2.9.0
passlib>=1.7.4
bcrypt>=4.0.1
jwt>=1.3.1
python-jose>=3.3.0
authlib>=1.2.1

# Database and Storage
sqlalchemy>=2.0.22
pymongo>=4.5.0
redis>=5.0.1
elasticsearch>=8.10.1
pinecone-client>=2.2.4
qdrant-client>=1.6.0
weaviate-client>=3.24.1
milvus>=2.3.0
clickhouse-driver>=0.2.6

# Monitoring and Logging
prometheus-client>=0.17.1
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
loguru>=0.7.2
rich>=13.6.0
tqdm>=4.66.1
sentry-sdk>=1.32.0

# Business Automation
zapier-python>=0.2.4
make-api-client>=0.1.0
n8n-python-client>=0.1.0
airtable-python-wrapper>=0.15.3
pygsheets>=2.0.6
trello>=0.9.7
slack-sdk>=3.23.0
tweepy>=4.14.0
facebook-sdk>=3.1.0
linkedin-api>=2.0.0
instagram-private-api>=1.6.0
youtube-api-python>=0.0.20

# Development and Testing
pytest>=7.4.2
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
pre-commit>=3.5.0
docker>=6.1.3
gunicorn>=21.2.0
uvicorn>=0.23.2
fastapi>=0.103.2
